+++ ismn/readers.py
@@ -1,34 +1,31 @@
-# Copyright (c) 2018, TU Wien, Department of Geodesy and Geoinformation
-# All rights reserved.
-
-# Redistribution and use in source and binary forms, with or without
-# modification, are permitted provided that the following conditions are met:
-#    * Redistributions of source code must retain the above copyright
-#      notice, this list of conditions and the following disclaimer.
-#    * Redistributions in binary form must reproduce the above copyright
-#      notice, this list of conditions and the following disclaimer in the
-#      documentation and/or other materials provided with the distribution.
-#    * Neither the name of TU Wien, Department of Geodesy and Geoinformation nor the
-#      names of its contributors may be used to endorse or promote products
-#      derived from this software without specific prior written permission.
-
-# THIS SOFTWARE IS PROVIDED BY THE COPYRIGHT HOLDERS AND CONTRIBUTORS "AS IS" AND
-# ANY EXPRESS OR IMPLIED WARRANTIES, INCLUDING, BUT NOT LIMITED TO, THE IMPLIED
-# WARRANTIES OF MERCHANTABILITY AND FITNESS FOR A PARTICULAR PURPOSE ARE
-# DISCLAIMED. IN NO EVENT SHALL THE COPYRIGHT HOLDER BE LIABLE FOR ANY
-# DIRECT, INDIRECT, INCIDENTAL, SPECIAL, EXEMPLARY, OR CONSEQUENTIAL DAMAGES
-# (INCLUDING, BUT NOT LIMITED TO, PROCUREMENT OF SUBSTITUTE GOODS OR SERVICES;
-# LOSS OF USE, DATA, OR PROFITS; OR BUSINESS INTERRUPTION) HOWEVER CAUSED AND
-# ON ANY THEORY OF LIABILITY, WHETHER IN CONTRACT, STRICT LIABILITY, OR TORT
-# (INCLUDING NEGLIGENCE OR OTHERWISE) ARISING IN ANY WAY OUT OF THE USE OF THIS
-# SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
+# The MIT License (MIT)
+#
+# Copyright (c) 2019 TU Wien
+#
+# Permission is hereby granted, free of charge, to any person obtaining a copy
+# of this software and associated documentation files (the "Software"), to deal
+# in the Software without restriction, including without limitation the rights
+# to use, copy, modify, merge, publish, distribute, sublicense, and/or sell
+# copies of the Software, and to permit persons to whom the Software is
+# furnished to do so, subject to the following conditions:
+#
+# The above copyright notice and this permission notice shall be included in all
+# copies or substantial portions of the Software.
+#
+# THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR
+# IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY,
+# FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE
+# AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER
+# LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM,
+# OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE
+# SOFTWARE.
 
 '''
 Created on Jul 31, 2013
 
 @author: Christoph Paulik
 
-Updated on Sep 23, 2018
+Updated on Dec 14, 2018
 
 @author: Philip Buttinger philip.buttinger@geo.tuwien.ac.at
 '''
@@ -38,6 +35,7 @@ import pandas as pd
 from datetime import datetime
 import numpy as np
 import logging
+import io
 
 
 variable_lookup = {'sm': 'soil moisture',
@@ -154,7 +152,7 @@ def get_info_from_file(filename):
     filename_elements : list
         filename without path split by _
     """
-    with open(filename, 'U') as f:
+    with io.open(filename, mode='r', newline=None) as f:
         header = f.readline()
     header_elements = header.split()
 
@@ -440,7 +438,7 @@ def get_min_max_timestamp_header_values(filename):
     """
     Get minimum and maximum observation timestamp from header values format.
     """
-    with open(filename, mode='rU') as fid:
+    with io.open(filename, mode='r', newline=None) as fid:
         _ = fid.readline()
         first = fid.readline()
         last = tail(fid)[0]
@@ -454,7 +452,7 @@ def get_min_max_timestamp_ceop_sep(filename):
     """
     Get minimum and maximum observation timestamp from ceop_sep format.
     """
-    with open(filename, mode='rU') as fid:
+    with io.open(filename, mode='r', newline=None) as fid:
         first = fid.readline()
         last = tail(fid)[0]
 
@@ -467,7 +465,7 @@ def get_min_max_timestamp_ceop(filename):
     """
     Get minimum and maximum observation timestamp from ceop format.
     """
-    with open(filename, mode='rU') as fid:
+    with io.open(filename, mode='r', newline=None) as fid:
         first = fid.readline()
         last = tail(fid)[0]